Moved tests to more appropriate, removed the issues helpers and moved them inline.
