+++ homeassistant/components/prowl/const.py
@@ -4,4 +4,3 @@ from homeassistant.const import Platform
 
 DOMAIN = "prowl"
 PLATFORMS = [Platform.NOTIFY]
-YAML_DEPRECATED_IN = "2026.1.0"
